hacky fix that gets us passing our tests on windows 7 with strawberry perl 5.22.0
we end u with "DIRFILESEP = \\" in our makefile on those systems, and the regular expressions in place to parse the Makefile can't handle that correctly. Here we 'manually fix' the things that breaks, until we can improve the parsing. This may not work everywhere, but it doesn't appear to break anything anywhere either.